Swapped class names, cleaned CSS & HTML, tidy'd.
Switched .exhibit to .tab and .tab to .exhibit. Changed project specific class name's to ID's and standardized the naming. Added comments to group the ID's in includes/style.css.

Got rid of empty li elements and moved virtually all style to the stylesheet as it should be.

For some reason I can't figure out, the #peepshow_entry_title-2 anchor element is not inheriting font-style:italics.