MySQL q's are now prepared. Added includes/cms.php
All MySQL queries from $_GET are now prepared. This logic has been placed in `$this->query($sql, $data_types, $dirty_data, $dirty_data...)`.

Projects are now also handled by includes/cms.php.

I still don't feel as though the code is entirely DRY, and the object schema still feels somewhat procedural.

The publish_* scripts need to be updated or gotten rid of for a nice interface.

The date_posted storage needs to be switched to datetime, that was a silly decision to start with.