Added cms.php
...Added cms.php, don't know how I missed that...

Everything works but #contact_me isn't spaced correctly and gets screwy with an exhibit open.

Have realized there is no way to erase what PHP has already printed to the client after issuing a 404. The method of output needs to be changed to allow for this.

Would very much like to implement note lists with /notes/YYYY/MM/DD.