Comment form works!
Still needs field validation, and the author/email fields need default values. The way the author/email is displayed above comments needs to be adjusted when there is no email to link to. The comments link on note/ pages should be changed when there are no comments yet.

Most importantly, the note class has gotten sloppy. $this->id is initiated from display_note(), creating a dependency for the form method. This should be moved into the constructor, but some thought should to be given to minimize database queries.

Also a few subtle formatting changes, but I still can't get the reCAPTCHA to align right without floating...