RSS feed working
Trusts note HTML, does not escape! This is acceptable, as only I will be editing the notes, but is worth mentioning again.

The first two sentences are found by locating periods, which is good enough I think.

Haven't added a link to the feed, honestly I'm not sure I want to, it's just nice to have the feature. I want to make sure the /notes/ page stays clean.

URL is /notes/rss