Comments now POSTed by AJAX.
Big problem now is that with JS disabled the CAPTCHA image doesn't load (although it IS fetched from google). I'm not sure why exactly that is. Originally the image was loaded with recaptcha_get_html($this->recaptcha_publickey), but that isn't compatible with the JS load.

Class captcha validates with google and returns the result.

The new comment is not loaded into view after submitting.

The fields should be validated as you go -> special note needs to be taken to consider the javascript-disabled validation and feedback.

Lots of prettying up is necessary with the error messages and such.

It's probably worth spending a commit on code cleanup now, it's getting a bit messy, particularly with the HEREDOCS.