sed -i s/\"/\'/g projects/* && rm print()s in pub*
This fixes a bug introduced in commit f9f5098 that broke the frontpage. The problem was that the method used to force the project page to show up searches for class='notes' but they were ""s in error. It's kind of a crude method in the first place, but it works.

The print statement in ./publish_projects was from debugging a while ago and served no purpose.