Fixed div error in index.php
Fixed error stopping the javascript exhibit display on the index page. The error was very strange (as in why the code was there, the fix was simple), I tried to track down its origin with blame, but git lost the history after a rename. This was already fixed in branch dev, but it was not a distinct commit. Still validating (at least the index page).